Separated DatePicker into another file, minor package.json change
